Moved arg parsing and user/host checking to each individual subsystem action. This is necessary to make it possible to call other scripts on other machines rather than MCP calling itself on another machine as another user. I wrapped as much as I could into wrapper methods to avoid cutting and pasting a bunch of code.

+++ mcp_modules/upload.py
@@ -15,7 +15,12 @@ class upload(subsystem):
             self.state['status']['page'] = "online"
 
     def lock_page(self, params):
-        self.parse_method_params('lock_page', [], {}, params)
+        action = 'lock_page'
+        self.parse_action_params(action, [], {}, params)
+        if self.check_userhost() == -1:
+            self.pass_mcp_cmd(action, params)
+            return 0
+
         fh = file(self.lock_file, 'a')
         try:
             os.utime(self.lock_file, None)
@@ -28,7 +33,12 @@ class upload(subsystem):
         f.write(jstate)
 
     def unlock_page(self, params):
-        self.parse_method_params('unlock_page', [], {}, params)
+        action = 'unlock_page'
+        self.parse_action_params(action, [], {}, params)
+        if self.check_userhost() == -1:
+            self.pass_mcp_cmd(action, params)
+            return 0
+
         if os.path.isfile(self.lock_file):
             os.unlink(self.lock_file)
